+++ plugins/gs/src/components/scaffolder/TemplateStringInput/index.ts
@@ -1,0 +1,2 @@
+export { TemplateStringInput } from './TemplateStringInput';
+export { TemplateStringInputSchema } from './schema';